last but one commit with final code

+++ update_script.sh
@@ -1,39 +1,51 @@
 #!/bin/sh
 environment=$1
-### clean up of the dynamic files ##
-rm finalfile.txt
 
-#echo "getting the commit id"
+if [ $# == 0 ]; then
+    echo "No environment details provided"
+    exit 1
+fi
+
+### clean up of the dynamic files before running the script##
+
+[ -f finalfile.txt ] && rm finalfile.txt
+[ -f targent_agent_list.txt ] && rm targent_agent_list.txt
+[ -f targent_agent.txt ] && rm targent_agent.txt
+
+
+#echo "Getting the commit id"
 commit_id=`git rev-parse HEAD`
-echo "listing the files with the commit id : $commit_id"
+echo "Below are the list of changed files with the commit id : $commit_id"
 `git show --pretty="" --name-only $commit_id > commit_id_files_list.txt`
-echo "file changes on this commit are : "
 cat commit_id_files_list.txt
 `cat mapping_files/$environment | cut -d ":" -f1 > searchagent_list.txt`
 commit_file='commit_id_files_list.txt'
 searchagent_list='searchagent_list.txt'
+
 i=1
-while read comment_file_name; do
-#Reading each comment_file_name.
-echo "$i file : $comment_file_name"
-while read sourceagentcomment_file_name; do
-cat $comment_file_name | grep $sourceagentcomment_file_name
-if [ $? == 0 ]
-then
-   #echo "we found the agent"
-   echo $sourceagentcomment_file_name >> targent_agent_list.txt
-   target_agent=`cat mapping_files/$environment | grep $sourceagentcomment_file_name | cut -d ":" -f2`
-   echo "target :$target_agent, source : $sourceagentcomment_file_name"
-   sed -e 's/Source_SecureAgent_Name/'$sourceagentcomment_file_name'/' -e 's/Target_SecureAgent_Name/'$target_agent'/' -e "s/comma/\,/" base_template >> finalfile.txt
-else
-   echo "we didnt find the $sourceagentcomment_file_name in $comment_file_name"
-fi
-done < $searchagent_list
-echo $?
+while read file_name; do
+echo "$i file : $file_name"
+	while read sourceline; do
+	cat $file_name | grep $sourceline
+	if [ $? == 0 ]
+	then
+	   echo "[INFO: YES, $sourceline in $file_name exists]"
+	   echo $sourceline >> targent_agent_list.txt
+	else
+	   echo "[INFO: No $sourceline in $file_name]"
+	fi
+	done < $searchagent_list
 i=$((i+1))
 done < $commit_file
 
 sort targent_agent_list.txt | uniq > targent_agent.txt
+
+while read sources_agent_name; do
+
+	target_agent=`cat mapping_files/$environment | grep $sources_agent_name | cut -d ":" -f2`
+	echo "source : $sources_agent_name --> Target :$target_agent"
+    sed -e 's/Source_SecureAgent_Name/'$sources_agent_name'/' -e 's/Target_SecureAgent_Name/'$target_agent'/' -e "s/comma/\,/" base_template >> finalfile.txt
+done < targent_agent.txt
+
 truncate -s -2 finalfile.txt
 sed '/#updatehere/r finalfile.txt' exisitingscript.txt > replacable_script.txt
-
